Update patch set 1
Patch Set 1: Code-Review-1

There is a mistake in the docker rm instructions that you added. The name of the 2 containers is wrong. Based on the docker run instructions, containers images are renamed as vfc_wf*, while docker rm has wf* (without vfc as prefix).

Could you please patch the contribution to match the correct name of the containers?

Thanks!

Patch-set: 1
Label: Code-Review=-1