Update patch set 1
Patch Set 1:

Hi Kanagaraj, as we agreed in Paris F2F, let's using CLI to register services in a separated script and it can be an alternative way for service registration after testing.
Currently, it's an urgent issue, I prefer we quickly fix it so we can move on to the integration test.

Patch-set: 1