Update patch set 1
Patch Set 1:

Hi Kanagaraj,
I believe you can fix them soon. But it might have other issues as well and I don't want to block the integration test. We could do it in parallel. After the CLI version is verified, we can use it for service registration.

This commit itself has no issue, so please +1 so it can be submitted.

For the usage of path, it's the customized publish path of the registered service.If path parameter is specified when registering the service, the service will be published to apigateway under this path. Otherwise, the service will be published to apigateway using a fixed format: api/{serviceName}/{version}. The customized publish path should only be used for back-compatible.

Patch-set: 1