Add direct designate zone creation code
Added DNS zone creation code for OpenStack installations with Designate
support.  Previously for this type of OpenStack instaalations we
assumed that there was no need to create new zone.  But such assumption
is not consistent with the DACE blueprints.  Therefore the blueprints
working for proxied Designate OpenStack installation will break not work
for direct Designate OpenStack installations. This new change makes the
vm_init behaves consistently for both direct Designate and proxied Designate
so that the same blueprint can be use in both environments.

Issue-ID: DCAEGEN2-200
Change-Id: I1f82f60facad77849c67f0feb1fa07f02af1079e
Signed-off-by: Lusheng Ji <lji@research.att.com>