Reset VES packet loss rate to zero on reboot
Adds logic to reset the VES packet loss rate to 0%
when the vg-mux is rebooted.

As part of this change, the VES agent configuration
will only be restored via the honeycomb persistence
mechanism.  So, if the VES agent is configured via
the vppctl CLI, that setting will not be saved during
shutdown.  In other words, setting of the VES agent
and VES mode should be done via the honeycomb REST
API in order for the configuration to be persisted
across reboot (with the pkt loss rate being reset to
0 %).

Change-Id: I8d1f43316d9608f8d7e7ce919f129ce8fe48aa7f
Issue-ID: INT-63
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>