delete vFW policy before updating resourceID
POLICY-486 deals with the issue that pushPolicy API
may not provision the latest version of the policy
when multiple versions are present.   While a fix
has been merged into amsterdam release, it has not
been released, the next 1.1.2 released version
has not been produced yet.

We met with Tarun to discuss the issue and suggested
to place a call to the deletePolicy in the script.
After a few tests, we think that this fix works
around the POLICY-486 issue, and seems that updates
the resourceID correctly, by ensuring that at there is
only 0 or 1 versions of the policy at a given time.

Change-Id: Ibcd20869d13dc115d57db5bc79b56513547324f7
Issue-ID: POLICY-486
Signed-off-by: Jorge Hernandez <jh1730@att.com>