Update patch set 1
Patch Set 1: Code-Review+1

Ok - I think I see what's happening.  The original vpp source file that is being patched by this patch has a lot of trailing white space, so hunk #6 of this patch is gets applied to the wrong place in the file because of those subtle white space differences.  (the patch applied with a fuzz of 1 - but compilation failed since the line was patched into the wrong spot in the source).

This patch puts adds some white space differences which allows that patch to apply without any fuzz.  I suppose this is fine.

An alternative fix would be to change the install script that invokes 'patch' to have it ignore whitespace - that works as well.  The hunks get patched in based on line #'s and context line matches.

But, I'm ok with this patch.

Patch-set: 1
Label: Code-Review=+1