Update patch set 1
Patch Set 1:

Helen, currently we are not supose to make any changes on this file to remove and add developers. 

This change is meant to reflect how the group permissions look right now. After this change is merged, we can start making changes to the file to remove or add users after Gilda's approval of course

Patch-set: 1