Update patch set 1
Patch Set 1: Code-Review-2 Verified-1

This is not necessary. The ONAP Heat template points to the Nexus raw repository, which contains install and init scripts. The 1.1.1 version is there and can be used as usual. 1.2.0 doesn't exist yet because the release isn't cut. 1.2.0-SNAPSHOT exists instead. For the VNFs, the situation is different. They still use the Nexus raw repository (this is why we set the version number to 1.2.0-SNAPSHOT, even though 1.1.1 was still good to use), but they also use the Nexus staging (or autorelease-xxx) repository, where the 1.1.1 jars and archives disappeared. This is why we have to use 1.2.0 for that (not that staging or autorelease-xxx don't use -SNAPSHOT suffix).

Patch-set: 1
Reviewer: Gerrit User 55 <55@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2
Label: Verified=-1