Update patch set 1
Patch Set 1: Code-Review-1

Regarding the issue with demo artifacts missing from staging:

It looks like the issue is with vnfs/vLBMS/apis/vlb-vnf-onap-distribution/pom.xml not inheriting from the root pom in the demo repo.  

Since it doesn’t inherit from the demo root pom, the maven command in the release job that strips the SNAPSHOT version from all pom.xml files misses this file, meaning that all poms get the version updated to non-SNAPSHOT except this one.  This means that this pom.xml still produces a SNAPSHOT artifact, which would cause a deploy to a staging repo to fail.

Is it possible to change this vnfs/vLBMS/apis/vlb-vnf-onap-distribution/pom.xml to inherit from the demo root pom instead of io.fd.honeycomb?  This would be the ideal setup.  If it’s not possible, we’ll have to think of a workaround to explicitly strip the SNAPSHOT version from this particular pom.

Patch-set: 1
Reviewer: Gerrit User 35 <35@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1