Refactor vCPE vG-MUX and vG image creation process
Separate heat/env templates and build scripts have been created for vG-MUX and vG

The build logic and parameters are removed from the install scripts and
the base heat/env files used to deploy the VMs.

For backwards compatibility, the compile_state parameter was added to both VMs. 

Small corrections were made to the heat templates for clarity.

Issue-ID: INT-615

Change-Id: I3180654f959776ebd4706613e00b9faff3f72e9e
Signed-off-by: Praneta Paithankar <praneta.paithankar@intel.com>