Update VBRG, vBNG compatibility w/ prev templates
The VNF heat templates and environment files for the Casablanca release
provide backwards compatibility with previous releases. In the build
heat template of each VNF, the compile_state parameter is included
with a default state of ‘done’ specified.

This change removes the compile_state parameter from .env and .yaml files
where it was not necessary in order to maintain consistency with the way
that backwards compatibility is provided in vG-MUX and vGW templates.

Issue-ID: INT-615

Change-Id: I6317f20a94c374d30edc019dfe36f0c28bca4830
Signed-off-by: Mariah Jacobs <mariah.jacobs@intel.com>