Do not prepand randomo str in key for vFW_NextGen
As the assignment for the key-name is using CDS, the key-name
shouldn't be adding anything to the name that has been resolved,
as it would create a discrepancy between what is expected to be
the key-name vs what it would be.
Using CDS, resolutions happens through GR-API, and are stored within
MD-SAL. If that value is altered, deployment will fail as the process
won't know about this additional random-str.

Change-Id: Icba069a1345572debded44b4e5486b23524728ea
Issue-ID: INT-694
Signed-off-by: Alexis de Talhouët <adetalhouet89@gmail.com>