Update patch set 1
Patch Set 1: Code-Review-1

Changes are fine, but are they required for Casablanca? The vFWCL is used for regression by Robot Framework. If these changes are required for Casablanca, then please work on Robot to make it consistent. Robot will be broken by these changes. If the changes are not strictly required for Casablanca, I would delay this contribution and accept it when the release branch is cut, so it won't impact Robot in the short term and we'll have enough time to update Robot itself for Dublin testing.

Marco

Patch-set: 1