Fix version of libevel code used for vgmux
The VES patch applied the vgmux code needs
to be applied against a version of the code
that was present near the date of Nov 20, 2017.
The code for libevel.so present in the repo at
that time produces events with the correct
attribute labels. (see refered Issue ID)

This patch provides a suitable commit id to use.
Replacing one which was not suitable.

Change-Id: I3460927e21ff23bc66ac77475f4d8d45921476e0
Issue-ID: TEST-132
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
(cherry picked from commit 95fd361eaff65980e6ff6f0e1e548d2822b0909b)