Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I ran it this same problem before. Could we do something at run time ?

I think that the best practice would be to contact clearwater project to see if it is a problem to reference new heat templates under Apache2

The heat templates of this patch were inspired by their initial community templates.
These templates have not changed since 2 years (except one file / 1 year), the are still referencing Ubuntu 14.04 base images for instance.

The current heat templates had been modified for ONAP in order to be fully VVP compliant. 

The consequence of a heat template under GPL v3 is not fully clear for me as heat templates are recipes consumed by Heat.
As it is a modeling template, it is difficult and weird to re-code it (structures are more or less imposed)

For me, it is similar to consider a web page/component versus the web server
there are lots of discussions on this topic, I just found a thread here: https://opensource.stackexchange.com/questions/4360/what-are-the-implications-of-licensing-a-javascript-library-under-gpl

"using a single GPL'd component on a website does not require open sourcing the whole website. But if you build your web app on top of some GPL'd component then the whole web app should be GPL'd as well."

The clearwater VNF is under GPL v3, to be deployed on an OpenStack (Apache 2) with possibly ONAP (apache v2) running on GNU Linux systems(GPL v2).
we understand why people from legal departments are so well paid :)

One solution could be to propose the evolution of the templates to clearwater and reference it in their repo under their licensing conditions and in our repo to indicate the URL where to find such templates.
The value is in the VNFs not really in the heat templates.

But in our case, I assume that the position is to take no risk and clean any GPL - or non apache 2 references in our repositories.

Patch-set: 1