Collectd Global CRD
This patch introduces an enhancement to the existing Collectd Operator.
It has now 2 CRDS CollectdGlobal and CollectdPlugin and corresponding
controllers. CollectdGlobal is used to specify the global options of
collectd conf and CollectdPlugin is used to load/unload plugins.

Added synchronization when both controllers try to fetch the same
resources. Refactored some of the common code to collectdutils.

Issue-ID: ONAPARC-461
Signed-off-by: Dileep Ranganathan <dileep.ranganathan@intel.com>
Change-Id: Id2edf6fa6de56d9d7216ebb258b3017661c6a3dd