Update patch set 5
Patch Set 5:

Most of it is fine because we do not use atm. But in fact some parameters could be used to populate data to AAI if we consider i.e. the day 2 configuration with APPC (Ansible, Netconf). But if we don't then we can leave it like that atm.

Patch-set: 5