Update patch set 5
Patch Set 5:

our current synchronization with AAI is very limited. Also the support of CNF in AAI is not the final one. But in order to use APPC configuration properly we need to have vserver and vnfc objects created with proper data. Most of the parameters removed (identifiers, naming code), if they were really assigned, could be used to populate this data to AAI somehow. But if we don't think about the APPC in the next release we can forget about it because CDS will take proper data from mdsal.

Patch-set: 5